Update README.md
The project if focused on creating a career service platform so that both students( or workers), and employers will be able to post and update their portfolio, and job postings respectively. To differentiate our platform with other competitors on the market like "indeed", and "Job bank" , or platform focuses on a intuitive and easy to use design.