Update README.md
The main purpose of this project is to create a career service platform, that allows both students(workers) , and employers to update their portfolio, and job postings respectively. Just like other similar websites, such as "indeed" and "Job bank", we strive to create and intuitive, simple to use and moder experience with our website.